requirements.txt
flask
python-docx
pypdf
requests
beautifulsoup4
openai
python-dotenv